+++ core/src/CteScore.ts
@@ -5,15 +5,15 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-import {IFn2Path, getNonce} from './CmnLib';
+import {IFn2Path, getNonce, chkBoolean} from './CmnLib';
 import {ScriptScanner} from './ScriptScanner';
 import {AnalyzeTagArg, HPRM} from './AnalyzeTagArg';
 
-import {CustomTextEditorProvider, TextDocument, WebviewPanel, CancellationToken, Uri, ExtensionContext, window, Webview, Range, WorkspaceEdit, workspace, Position} from 'vscode';
+import {TextDocument, WebviewPanel, CancellationToken, Uri, ExtensionContext, window, Webview, Range, WorkspaceEdit, workspace, Position} from 'vscode';
 import fs = require('fs-extra');
 
 interface HTDS {
-	col?		: number,
+	col?		: string,
 	td_style?	: string,
 	btn_style?	: string,
 	icon		: string,
@@ -21,19 +21,34 @@ interface HTDS {
 	tooltip?	: string,
 	args?		: {
 		name	: string;
-		type	: 'bool'|'str'|'num'|'rule'|'textarea';
+		type	: 'bool'|'str'|'num'|'textarea'|'select'|'range';
 		val?	: string;
 		hint?	: string;
+		key?	: string;
+		exts?	: string;
+		filter?	: string;
+		min?	: string;
+		max?	: string;
+		step?	: string;
 	}[],
 	detail?		: string,
 }
 
-export class CteScore implements CustomTextEditorProvider {
-	private	static	htmBaseSrc	= '';
+export class CteScore {
 	private	static	localExtensionResRoots: Uri;
+	private	static	htmBaseSrc	= '';
 	static	init(ctx: ExtensionContext): void {
-		const cteScr = new CteScore;
-		window.registerCustomEditorProvider('SKYNovel.score', cteScr);
+		window.registerCustomEditorProvider('SKYNovel.score', {
+			async resolveCustomTextEditor(doc: TextDocument, webviewPanel: WebviewPanel, _token: CancellationToken): Promise<void> {
+				const path = doc.fileName;
+				for (const cur in CteScore.hCur2Me) {
+					if (cur === path.slice(0, cur.length)) {
+						CteScore.hCur2Me[cur].resolveCustomTextEditor(doc, webviewPanel);
+						break;
+					}
+				}
+			}
+		});
 
 		const path_ext_res = ctx.extensionPath +`/res/webview/`;
 		CteScore.localExtensionResRoots = Uri.file(path_ext_res);
@@ -46,14 +61,17 @@ export class CteScore implements CustomTextEditorProvider {
 		.replace(/<div class="card-group">[\s\S]+<\/div><!-- card-group  -->/, '<div class="card-group"> </div><!-- card-group  -->');
 	}
 
-	private	static	hPath2Tokens	: {[path: string]: {
+	private	static	hCur2Me		: {[path: string]: CteScore}	= {};
+	constructor(readonly curPrj: string) {CteScore.hCur2Me[curPrj] = this;}
+
+	private	hPath2Tokens	: {[path: string]: {
 		uriPrj	: Uri;
 		htm		: string;
 		skipupd	: boolean;
 	}}	= {};
 	updScore(path: string, curPrj: string, aToken: string[]) {
 		const s = {line: 1};
-		CteScore.hPath2Tokens[path] = {
+		this.hPath2Tokens[path] = {
 			uriPrj	: Uri.parse(curPrj),
 			htm		: CteScore.htmBaseSrc.replace('<tbody/>', `<tbody>${
 				aToken.map((token, i)=> this.token2html(s, token, i)).join('')
@@ -61,8 +79,8 @@ export class CteScore implements CustomTextEditorProvider {
 			skipupd	: false,
 		};
 
-		const wv = CteScore.hPath2Wv[path];
-		if (wv) wv.html = this.repWvUri(CteScore.hPath2Tokens[path].htm, wv);
+		const wv = this.hPath2Wv[path];
+		if (wv) wv.html = this.repWvUri(this.hPath2Tokens[path].htm, wv);
 	}
 	private	repWvUri(inp: string, wv: Webview): string {
 		return inp
@@ -71,20 +89,20 @@ export class CteScore implements CustomTextEditorProvider {
 	}
 
 	isSkipUpd(path: string): boolean {
-		const t = CteScore.hPath2Tokens[path];
+		const t = this.hPath2Tokens[path];
 		if (t?.skipupd) {t.skipupd = false; return true;}
 		return false;
 	}
 
 
 	separation(path: string) {	// 分離
-		CteScore.hPath2Wv[path]?.postMessage({cmd: 'separation'});
+		this.hPath2Wv[path]?.postMessage({cmd: 'separation'});
 	}
 	combining(path: string) {	// 結合
-		CteScore.hPath2Wv[path]?.postMessage({cmd: 'combining'});
+		this.hPath2Wv[path]?.postMessage({cmd: 'combining'});
 	}
 	updLine(doc: TextDocument, rng: Range, txt: string, aToken: string[]): boolean {	// 分離状態でコールすること
-		const wv = CteScore.hPath2Wv[doc.fileName];
+		const wv = this.hPath2Wv[doc.fileName];
 		if (! wv) return false;
 
 		const sl = rng.start.line;
@@ -93,11 +111,11 @@ export class CteScore implements CustomTextEditorProvider {
 			rng.end.character > 0) && sl !== el) return true;
 			// 全スキャン。せっかくScoreなので差分更新はオーバースペック
 
-		if (txt === '') {wv.postMessage({cmd: 'del', sl: sl}); return false;}
+		if (txt === '') {wv.postMessage({cmd: 'del', ln: sl}); return false;}
 
 		wv.postMessage({
 			cmd	: (sl === el && txt.slice(-1) === '\n') ?'ins' :'rep',
-			sl	: sl,
+			ln	: sl,
 			htm	: (txt === '\n'
 				? this.token2html({line: sl +1}, '\n\n', -1)
 				: this.repWvUri(aToken.map(t=> (t.charCodeAt(0) < 11)
@@ -109,22 +127,12 @@ export class CteScore implements CustomTextEditorProvider {
 		return false;
 	}
 
-	undefMacro(def_nm: string) {delete CteScore.hTag2Tds[def_nm];}
-	defMacro(def_nm: string, hPrm: any) {
-//console.log(`fn:CteScore.ts line:114 defMacro def_nm:${def_nm} hPrm:${JSON.stringify(hPrm)}`);
-		CteScore.hTag2Tds[def_nm] = ()=> {return {...{
-			icon		: 'icon未指定',
-			btn_face	: 'btn_face未指定',
-		}, ...hPrm}};
-	}
 
-
-	private	static	hPath2Wv	: {[path: string]: Webview}	= {};
-	async resolveCustomTextEditor(doc: TextDocument, webviewPanel: WebviewPanel, _token: CancellationToken): Promise<void> {
-		const t = CteScore.hPath2Tokens[doc.fileName];
-		const wv = webviewPanel.webview;
+	private	hPath2Wv	: {[path: string]: Webview}	= {};
+	private	resolveCustomTextEditor(doc: TextDocument, webviewPanel: WebviewPanel) {
 		const path = doc.fileName;
-		CteScore.hPath2Wv[path] = wv;
+		const t = this.hPath2Tokens[path];
+		const wv = this.hPath2Wv[path] = webviewPanel.webview;
 		wv.options = {
 			enableScripts: true,
 			localResourceRoots: [CteScore.localExtensionResRoots, t.uriPrj],
@@ -133,10 +141,11 @@ export class CteScore implements CustomTextEditorProvider {
 			switch (o.cmd) {
 			case 'info':	window.showInformationMessage(o.text); break;
 			case 'warn':	window.showWarningMessage(o.text); break;
+			case 'err':		window.showErrorMessage(o.text); break;
 
-			case 'loaded':	this.updWv_db(path, true);	break;
+			case 'loaded':	this.updWv_db(path); this.combining(path);	break;
 			case 'save_tbody':
-				CteScore.hPath2Tokens[path].htm = wv.html
+				this.hPath2Tokens[path].htm = wv.html
 				.replace(/<tbody>[\s\S]+<\/tbody>/,`<tbody>${o.tbody}</tbody>`);
 				break;
 
@@ -160,19 +169,25 @@ export class CteScore implements CustomTextEditorProvider {
 			}
 				break;
 
-			case 'add_req':{
+			case 'tool_put':{
+				const scr = String(o.scr)
+				.replace(/\$\{(.+)\}/g, (_, p1)=> {
+					let ret = '';
+					for (const i of this.hBufWords[p1]) {ret = `#${i}#`; break;}
+					return ret;
+				});
+console.log(`fn:CteScore.ts line:176 tool_put row:${o.row} to:${o.to} scr=${scr}=`);
 				wv.postMessage({
-					cmd	: 'add_res',
+					cmd	: 'tool_res',
 					row	: o.row,
-					htm	: this.repWvUri(this.token2html({line: o.row}, o.scr, o.row), wv),
+					htm	: this.repWvUri(this.token2html({line: o.row}, scr, o.row), wv),
 				});
 
 				t.skipupd = true;
 				const ed = new WorkspaceEdit();
-				ed.insert(doc.uri, new Position(o.to, 0), (o.scr === '\n\n') ?'\n' :(o.scr +'\n'));
+				ed.insert(doc.uri, new Position(o.to, 0), (scr === '\n\n') ?'\n' :(scr +'\n'));
 				workspace.applyEdit(ed);
-			}
-				break;
+			}	break;
 
 			case 'del':{
 				t.skipupd = true;
@@ -183,23 +198,49 @@ export class CteScore implements CustomTextEditorProvider {
 				break;
 
 			case 'input':{
-//console.log(`fn:CteScore.ts line:172 input lnum:${o.lnum} nm:${o.nm} val:${o.val}`);
+/**/console.log(`fn:CteScore.ts line:184 input ln:${o.ln} nm=${o.nm}= val=${o.val}=`);
 				t.skipupd = true;
 				const ed = new WorkspaceEdit();
-				const rng = new Range(o.lnum, 0, o.lnum +1, 0);
-				const txt = doc.getText(rng);
-				ed.replace(
-					doc.uri,
-					new Range(o.lnum, 0, o.lnum +1, 0),
-					txt.replace(/(\S+=#).*#/, `$1${o.val}#`)
+				const rng = new Range(o.ln, 0, o.ln +1, 0);
+				/*
+					case 38:	// & 変数操作・変数表示
+					case 42:	// * ラベル
+					case 59:	// ; コメント
+					case 91:	// [ タグ開始
+				*/
+				const base = (o.nm.charCodeAt(0) < 60)
+					? o.nm + o.val +'\n'
+					: doc.getText(rng);
+				let txt = base.replace(
+					new RegExp(`(${o.nm}=)([#"']).*\\2`),
+					`$1$2${o.val}$2`
+				).replace(
+					new RegExp(`(${o.nm}=)[^\\] #"']+`),
+					`$1${o.val}`
 				);
+				if (base === txt) txt = txt.replace(/(])/, ` ${o.nm}=#${o.val}#$1`);	// 属性追加
+				ed.replace(doc.uri, rng, txt);
 				workspace.applyEdit(ed);
-			}
+
+				const a_tag: any = ScriptScanner.analyzTagArg(txt);
+				const g = a_tag?.groups;
+				if (! g) break;
+				const t2t = this.hTag2Tds[g.name];
+				if (! t2t) break;
+				CteScore.alzTagArg.go(g.args);
+				const oTds = t2t(CteScore.alzTagArg.hPrm);
+				wv.postMessage({cmd: 'upd_btn_face', ln: o.ln, htm: `
+<i class="fas ${oTds.icon}" aria-hidden="true"></i>
+${oTds.btn_face}`, td: `<td class="p-0 ${oTds.td_style}">`, nm: o.nm, val: o.val});
+			}	break;
+
+			case 'req_wds':
+				wv.postMessage({cmd: 'res_wds', key: o.key, aWd: [...this.hBufWords[o.key].values()]});
 				break;
 			}
 		}, false);
 
-		wv.html = this.repWvUri(CteScore.hPath2Tokens[doc.uri.path].htm, wv);
+		wv.html = this.repWvUri(this.hPath2Tokens[doc.uri.path].htm, wv);
 
 		// 空ファイルなら適当なテンプレを挿入
 		if (doc.getText(new Range(0, 0, 1, 1)) === '') {
@@ -210,14 +251,14 @@ export class CteScore implements CustomTextEditorProvider {
 		// NOTE: sn→ssnらしきものに強制改行を入れるのならここでか
 	}
 	private	static	regFld	= /\w+/;
-	private	updWv_db(path: string, combining: boolean) {
+	private	updWv_db(path: string) {
 		const hFld2: {[fld: string]: {
 			ext		: string;
 			path	: string;
 			fn		: string;
 		}[]} =	{};
-		const t = CteScore.hPath2Tokens[path];
-		const hPath = CteScore.hPrj2hPath[t.uriPrj.path];
+		const t = this.hPath2Tokens[path];
+		const hPath = this.hPrj2hPath;
 		for (const fn in hPath) {
 			const p = hPath[fn];
 			for (const ext in p) {
@@ -234,23 +275,28 @@ export class CteScore implements CustomTextEditorProvider {
 			}
 		}
 
-		const wv = CteScore.hPath2Wv[path];
+		const wv = this.hPath2Wv[path];
 		wv.postMessage({
 			cmd			: 'upd_db',
 			pathPrj		: String(wv.asWebviewUri(t.uriPrj)),	// 最後に「/」必要
 			hFld2url	: hFld2,
 			hPath		: hPath,
-			combining	: combining,
 		});
 	}
 
+	private	hBufWords	: {[key: string]: Set<string>}	= {};
+	updWords(key: string, Words: Set<string>) {
+		this.hBufWords[key] = Words;
+		for (const path in this.hPath2Wv) {
+			this.hPath2Wv[path].postMessage({cmd: 'del_wds', key: key});
+		}
+	}
 
-	private	static	hPrj2hPath	: {[prj: string]: IFn2Path}	= {};
-	updPath(curPrj: string, hPath: IFn2Path) {
-		CteScore.hPrj2hPath[curPrj] = hPath;
-		for (const path_doc in CteScore.hPath2Tokens) {
-			if (path_doc in CteScore.hPath2Wv
-			&&	CteScore.hPath2Tokens[path_doc].uriPrj.path === curPrj) this.updWv_db(path_doc, false);
+	private	hPrj2hPath	: IFn2Path	= {};
+	updPath(hPath: IFn2Path) {
+		this.hPrj2hPath = hPath;
+		for (const path_doc in this.hPath2Tokens) {
+			if (path_doc in this.hPath2Wv) this.updWv_db(path_doc);
 		}
 	}
 
@@ -271,36 +317,43 @@ export class CteScore implements CustomTextEditorProvider {
 				return str_r;
 
 			case 38:	// & 変数操作・変数表示
+				// NOTE: [let]があるので不要かも
 				const is_dsp = (token.slice(-1) === '&');
 				tds = this.make_tds(
-					0, stt.line, 'btn-secondary" data-face="true',
+					0, stt.line, 'btn-secondary dropdown-toggle" data-face="true" data-mdb-toggle="dropdown" aria-expanded="false',
 					'fa-calculator',
 					is_dsp
 						? '変数表示'
 						: `変数操作 ${
 							token.slice(1, token.indexOf('=') +2)
 						}…`,
-					is_dsp ?token.slice(1, -1) :token.slice(1),
-				);
+					is_dsp ?token.slice(1, -1) :token.slice(1), `
+	<div class="form-outline col-12">
+		<input type="text" data-nm="&" id="${idx}txf" class="form-control" value="${token.slice(1)}"/>
+		<label class="form-label" for="${idx}txf">${
+			is_dsp ?'変数表示' :'変数操作'
+		}</label>
+	</div>`		);
 				break;
 
 			case 59:	// ; コメント
 				tds = this.make_tds(
 					0, stt.line, 'btn-outline-light btn-rounded dropdown-toggle" data-face="true" data-mdb-toggle="dropdown" aria-expanded="false',
-					'fa-comment-dots',
-					token.slice(1, 11) +'…',
-					token.slice(1), `
+					'fa-comment-dots', token.slice(1, 11)+'…', token.slice(1), `
 	<div class="form-outline col-12">
-		<textarea class="form-control" placeholder="コメントを入力" cols="40" rows="2" data-nm="text" id="sn-cm${idx}">${token}</textarea>
-		<label class="form-label" for="sn-cm${idx}">コメント</label>
+		<input type="text" data-nm=";" id="${idx}txf" class="form-control" value="${token.slice(1)}"/>
+		<label class="form-label" for="${idx}txf">コメント</label>
 	</div>`		);
 				break;
 
 			case 42:	// * ラベル
 				tds = this.make_tds(
-					0, stt.line, 'btn-outline-light" data-face="true',
-					'fa-tag', token.slice(1),
-				);
+					0, stt.line, 'btn-outline-light dropdown-toggle" data-mdb-toggle="dropdown" aria-expanded="false" data-face="true',
+					'fa-tag', token.slice(1), token.slice(1), `
+	<div class="form-outline col-12">
+		<input type="text" data-nm="*" id="${idx}txf" class="form-control" value="${token.slice(1)}"/>
+		<label class="form-label" for="${idx}txf">ラベル</label>
+	</div>`		);
 				break;
 
 			case 91:	// [ タグ開始
@@ -317,8 +370,8 @@ export class CteScore implements CustomTextEditorProvider {
 					5, stt.line, 'btn-outline-primary text-white dropdown-toggle sn-ext_txt" data-face="true" data-mdb-toggle="dropdown" aria-expanded="false',
 					'fa-align-left', '', '', `
 	<div class="form-outline col-12">
-		<textarea class="form-control" placeholder="本文テキストを入力" cols="40" rows="2" data-nm="text" id="sn-ta${idx}">${token}</textarea>
-		<label class="form-label" for="sn-ta${idx}">本文テキスト</label>
+		<textarea class="form-control" placeholder="本文テキストを入力" cols="40" rows="2" data-nm="text" id="${idx}ta">${token}</textarea>
+		<label class="form-label" for="${idx}ta">本文テキスト</label>
 	</div>`		);
 		}
 
@@ -341,7 +394,7 @@ export class CteScore implements CustomTextEditorProvider {
 		const g = a_tag?.groups;
 		if (! g) return '';
 
-		const t2t = CteScore.hTag2Tds[g.name];
+		const t2t = this.hTag2Tds[g.name];
 		if (! t2t) return this.make_tds(
 			0, line, 'btn-light',
 			'fa-code', g.name, g.args,
@@ -355,65 +408,69 @@ export class CteScore implements CustomTextEditorProvider {
 		:	len === 2 ?'col-2'
 		:	'';
 		return this.make_tds(
-			oTds.col ?? 0, line,
+			oTds.col ?this.hCn2Col[oTds.col] :0, line,
 			(oTds.btn_style ?? 'btn-secondary')
 			+ (oTds.args ?` dropdown-toggle" data-mdb-toggle="dropdown" aria-expanded="false` :''),
 			oTds.icon, oTds.btn_face,
 			oTds.tooltip ?? g.args,
-			oTds.args ?oTds.args.map(v=> {
+			oTds.args ?oTds.args.map((v, i)=> {
 				let ret = '';
-				let fo = 'form-outline';
+				let fo = 'form-outline px-2';
+				const id = `sn-${row}:${i}`;
 				switch (v.type) {
-					case 'bool':	ret = `
-<div class="form-check ml-3 py-2 ${frm_style}">
-<input type="checkbox" value="${v.val}" class="form-check-input px-0" id="sn-chk${row}" checked/>
-<label class="form-check-label" for="sn-chk${row}">${v.hint ?? v.name}</label>
-</div>`;					break;
-
-					case 'rule':	ret = `
-<div class="input-group mb-3 ${frm_style}">
-<button type="button" class="btn btn-success btn-sm px-2 text-start text-lowercase sn-fld" id="sn-btn${row}" data-ripple-color="dark" data-target="#sn-grpModal" data-title="背景選択" data-nm="rule">
-	<i class="fas fa-image"></i>
-	${v.hint ?? v.name}
-</button>
-</div>`;
-/*
-<input type="text" value="${v.val}" class="form-control" placeholder="${v.hint ?? v.name}" aria-label="${v.hint ?? v.name}" aria-describedby="sn-btn${row}" readonly/>
-
-
-					case 'rule':	ret = `
-<div class="input-group mb-3 ${frm_style}">
-<button type="button" class="btn btn-success btn-sm px-2 text-start text-lowercase sn-fld" id="sn-btn${row}" data-ripple-color="dark" data-mdb-toggle="modal" data-target="#sn-grpModal" data-title="背景選択" data-nm="rule">
-	<i class="fas fa-image"></i>
-	${v.hint ?? v.name}
-</button>
-<input type="text" value="${v.val}" class="form-control" placeholder="${v.hint ?? v.name}" aria-label="${v.hint ?? v.name}" aria-describedby="sn-btn${row}" readonly/>
-</div>`;
-*/
-						fo = '';
-						// form-outlineを使うとエラー
-						// しかもhtm出力して動かさないとわからない
-						break;
+					case 'bool':
+						fo = 'p-0';	// form-outlineを使うとエラー
+						ret = `
+<div class="form-control ${frm_style}">
+<div class="form-check">
+	<input type="checkbox" id="${id}chk" class="form-check-input" data-nm="${v.name}" value="${v.val}"${chkBoolean(v.val) ?' checked' :''}/>
+	<label class="form-check-label" for="${id}chk">${v.hint ?? v.name}</label>
+</div>
+</div>`;				break;
 
 					case 'textarea':
 						ret = `
-<textarea class="form-control" placeholder="${v.hint ?? v.name}を入力" cols="40" rows="2" data-nm="${v.name}" id="sn-ta${row}">${v.val}</textarea>
-<label class="form-label" for="sn-ta${row}">${v.hint ?? v.name}</label>`;
+<textarea id="${id}ta" class="form-control" placeholder="${v.hint ?? v.name}を入力" cols="40" rows="2" data-nm="${v.name}">${v.val}</textarea>
+<label class="form-label" for="${id}ta">${v.hint ?? v.name}</label>`;
 						break;
+
+					case 'select':
+						fo = 'p-0';	// form-outlineを使うとエラー
+						ret = `
+<div class="form-control">
+<label class="form-label select-label" for="${id}sel">${v.hint ?? v.name}</label>
+<select class="select" id="${id}sel" data-nm="${v.name}" data-key="${v.key}"${
+	v.exts ?`data-exts="${v.exts}"` :''
+}${
+	v.filter ?`data-filter="${v.filter}"` :''
+}>
+	<option selected>${v.val}</option>
+	<option>...</option>
+	<option>...</option>
+</select>
+</div>`;				break;
+
+					case 'range':
+						fo = 'p-0';	// form-outlineを使うとエラー
+						ret = `
+<div class="range form-control">
+	<label class="form-label" for="${id}rng">${v.hint ?? v.name}（${v.min} - ${v.max}）</label>
+	<input type="range" id="${id}rng" data-nm="${v.name}" class="form-range" min=${v.min} max=${v.max} step=${v.step} value="${v.val}"/>
+</div>`;				break;
 
 					default:	ret = `
 <input type="${
 v.type === 'num' ?'number' :'text'
-}" value="${v.val}" id="sn-txf${row}" class="form-control ${frm_style}"/>
-<label class="form-label" for="sn-txf${row}">${v.hint ?? v.name}</label>`;
+}" id="${id}txf" class="form-control ${frm_style}" data-nm="${v.name}" value="${v.val}"/>
+<label class="form-label" for="${id}txf">${v.hint ?? v.name}</label>`;
 				}
 				return `
 <div class="${fo} col-12 ${
 len === 1	? ''
 : len === 2	? 'col-md-6'
 : 'col-md-6 col-lg-4'
-} mt-3 ${
-v.type === 'bool' ?'pt-2' :''
+} my-1${
+v.type === 'bool' ?' pt-2' :''
 }">${ret}</div>`;
 			}).join('') :'',
 			oTds.td_style ?? '',
@@ -421,10 +478,19 @@ v.type === 'bool' ?'pt-2' :''
 		);
 	}
 
+
+	undefMacro(def_nm: string) {delete this.hTag2Tds[def_nm];}
+	defMacro(def_nm: string, hPrm: any) {
+		this.hTag2Tds[def_nm] = ()=> {return {...{
+			icon		: 'icon未指定',
+			btn_face	: 'btn_face未指定',
+		}, ...hPrm}};
+	}
 	private	static	readonly	alzTagArg	= new AnalyzeTagArg;
-	private	static	hTag2Tds	: {[tag_nm: string]: (hPrm: HPRM)=> HTDS}	= {
+	private	static	readonly	EXTS_PIC	= 'jpg|jpeg|png';
+	private	hTag2Tds	: {[tag_nm: string]: (hPrm: HPRM)=> HTDS}	= {
 		ch			: hPrm=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-outline-primary text-white dropdown-toggle sn-ext_txt" data-face="true" data-mdb-toggle="dropdown" aria-expanded="false',
 			icon		: 'fa-align-left',
 			btn_face	: '',
@@ -452,18 +518,9 @@ v.type === 'bool' ?'pt-2' :''
 			btn_face	: '戻る',
 		}},
 		add_lay		: hPrm=> {
-			let col = 0;
-			switch (hPrm.layer?.val) {
-				case 'base': col = 1; break;
-				case '0': col = 2; break;
-				case '1': col = 3; break;
-				case '2': col = 4; break;
-				case 'mes': col = 5; break;
-				case 'mes_c2p': col = 6; break;
-			}
 			const is_grp = (hPrm.class?.val === 'grp');
 			return {
-				col			: col,
+				col			: hPrm.layer?.val ?? 'base',
 				btn_style	: `btn-${
 					is_grp ?'success' :'primary text-white'
 				}`,
@@ -472,67 +529,72 @@ v.type === 'bool' ?'pt-2' :''
 					+ hPrm.layer?.val,
 			};
 		},
-		clear_lay	: hPrm=> {
-			let col = 0;
-			switch (hPrm.layer?.val) {
-				case 'base': col = 1; break;
-				case '0': col = 2; break;
-				case '1': col = 3; break;
-				case '2': col = 4; break;
-				case 'mes': col = 5; break;
-				case 'mes_c2p': col = 6; break;
-			}
-			return {
-				col			: col,
-				btn_style	: `btn-outline-primary text-white`,
-				icon		: 'fa-ban',
-				btn_face	: 'レイヤ設定消去',
-			};
-		},
-		lay			: hPrm=> {
-			let col = 0;
-			switch (hPrm.layer?.val) {
-				case 'base': col = 1; break;
-				case '0': col = 2; break;
-				case '1': col = 3; break;
-				case '2': col = 4; break;
-				case 'mes': col = 5; break;
-				case 'mes_c2p': col = 6; break;
-			}
-			return {
-				col			: col,
-				btn_style	: `btn-outline-primary text-white`,
-				icon		: 'fa-cog',
-				btn_face	: 'レイヤ設定',
-			};
-		},
+		clear_lay	: hPrm=> {return {
+			col			: hPrm.layer?.val ?? 'base',
+			btn_style	: `btn-outline-primary text-white`,
+			icon		: 'fa-ban',
+			btn_face	: 'レイヤ設定消去',
+		};},
+		lay			: hPrm=> {return {
+			col			: hPrm.layer?.val ?? 'base',
+			btn_style	: `btn-outline-primary text-white`,
+			icon		: 'fa-cog',
+			btn_face	: 'レイヤ設定',
+		};},
 
 		img			: hPrm=> {
-			let col = 0;
-			switch (hPrm.layer?.val) {
-				case 'base': col = 1; break;
-				case '0': col = 2; break;
-				case '1': col = 3; break;
-				case '2': col = 4; break;
-				case 'mes': col = 5; break;
-				case 'mes_c2p': col = 6; break;
-			}
+			const layer = hPrm.layer?.val ?? 'base';
 			return {
-				col			: col,
+				col			: layer,
 				btn_style	: `btn${
 					hPrm.fn?.val ?'' :'-outline'
 				}-${
-					col < 5
+					layer !== 'mes'	// TODO: 雑。文字レイヤかどうかで
 						? 'success'
 						: 'secondary text-white'
 				}`,
-				icon		: (col === 1) ?'fa-image' :'fa-user',
-				btn_face	: hPrm.fn?.val ?? ((col === 1) ?'(消去)' :'(退場)'),
+				icon		: (layer === 'base') ?'fa-image' :'fa-user',
+				btn_face	: hPrm.fn?.val ?? ((layer === 'base') ?'(消去)' :'(退場)'),
+				args		: [
+					{name: 'fn', type: 'select', val: hPrm.fn?.val ?? '(消去)', hint: '画像名', key: '画像ファイル名'},
+				]
 			};
 		},
+		scene_change	: hPrm=> {return {
+			col			: 'base',
+			td_style	: 'sn-cmb-'
+				+ (hPrm.bg?.val ?`start" data-fn="${hPrm.bg?.val}` :'end'),
+			btn_style	: 'btn-success text-black',
+			icon		: 'fa-door-open',
+			btn_face	: '場面転換'
+				+ (hPrm.rule?.val ?` ${hPrm.rule?.val}ルールで` :'')
+				+ (hPrm.time?.val ?` ${hPrm.time.val}msで` :'')
+				+ (` ${hPrm.bg?.val}へ変更` ?? '(消去)'),
+			args		: [
+				{name: 'bg', type: 'select', val: hPrm.bg?.val ?? '(消去)', hint: '背景の画像名', key: '画像ファイル名', exts: CteScore.EXTS_PIC, filter: 'bg\/'},
+				{name: 'time', type: 'num', val: hPrm.time?.val ?? '1000', hint: 'かける時間'},
+				{name: 'rule', type: 'select', val: hPrm.rule?.val ?? '', hint: 'ルール画像名', key: '画像ファイル名', exts: CteScore.EXTS_PIC, filter: 'rule\/'},
+			],
+		};},
+		scene_reserve	: hPrm=> {return {
+			col			: hPrm.layer?.val ?? '0',
+			td_style	: 'sn-cmb-'
+				+ (hPrm.fn?.val ?`start" data-fn="${hPrm.fn?.val}` :'end'),
+			btn_style	: 'btn-success btn-rounded text-black',
+			icon		: 'fa-user-ninja',
+			btn_face	: `場面準備 ${hPrm.fn?.val}`,
+			args		: [
+				{name: 'layer', type: 'select', val: hPrm.layer?.val ?? '0', hint: '画像レイヤ', key: '画像レイヤ名'},
+				{name: 'fn', type: 'select', val: hPrm.fn?.val ?? '(消去)', hint: '画像名', key: '画像ファイル名'},
+				{name: 'face', type: 'str', val: hPrm.face?.val ?? '', hint: '差分'},
+				{name: 'pos', type: 'str', val: hPrm.pos?.val ?? '', hint: '表示位置'},
+				{name: 'alpha', type: 'range', val: hPrm.alpha?.val ?? '', hint: '不透明度',  min: '0.0', max: '1.0', step: '0.1'},
+			],
+		};},
+
 		grp			: hPrm=> {
 			return {
-				col			: 1,
+				col			: 'base',
 				td_style	: 'sn-cmb-'
 					+ (hPrm.bg?.val ?`start" data-fn="${hPrm.bg?.val}` :'end'),
 				btn_style	: 'btn-success',
@@ -541,21 +603,21 @@ v.type === 'bool' ?'pt-2' :''
 					+ (hPrm.rule?.val ?` ${hPrm.rule?.val}ルールで` :'')
 					+ (hPrm.time?.val ?` へ変更 ${hPrm.time.val}msで` :''),
 				args		: [
-					{name: 'bg', type: 'str', val: hPrm.bg?.val ?? '(消去)', hint: '背景の画像名'},
+					{name: 'bg', type: 'select', val: hPrm.bg?.val ?? '(消去)', hint: '背景の画像名', key: '画像ファイル名', exts: CteScore.EXTS_PIC, filter: 'bg\/'},
 					{name: 'time', type: 'num', val: hPrm.time?.val ?? '1000', hint: 'かける時間'},
-					{name: 'rule', type: 'rule', val: hPrm.rule?.val ?? '', hint: 'ルール画像名'},
+					{name: 'rule', type: 'select', val: hPrm.rule?.val ?? '', hint: 'ルール画像名', key: '画像ファイル名', exts: CteScore.EXTS_PIC, filter: 'rule\/'},
 
-					{name: 'l0', type: 'str', val: hPrm.l0?.val ?? '(消去)', hint: 'レイヤ0の画像名'},
+					{name: 'l0', type: 'select', val: hPrm.l0?.val ?? '(消去)', hint: 'レイヤ0の画像名', key: '画像ファイル名'},
 					{name: 'f0', type: 'str', val: hPrm.f0?.val ?? '', hint: 'レイヤ0のface属性'},
 					{name: 'pos0', type: 'str', val: hPrm.pos0?.val ?? '', hint: 'レイヤ0のpos位置'},
 					{name: 'o0', type: 'num', val: hPrm.o0?.val ?? '', hint: 'レイヤ0の不透明度'},
 
-					{name: 'l1', type: 'str', val: hPrm.l1?.val ?? '(消去)', hint: 'レイヤ1の画像名'},
+					{name: 'l1', type: 'select', val: hPrm.l1?.val ?? '(消去)', hint: 'レイヤ1の画像名', key: '画像ファイル名'},
 					{name: 'f1', type: 'str', val: hPrm.f1?.val ?? '', hint: 'レイヤ1のface属性'},
 					{name: 'pos1', type: 'str', val: hPrm.pos1?.val ?? '', hint: 'レイヤ1のpos位置'},
 					{name: 'o1', type: 'num', val: hPrm.o1?.val ?? '', hint: 'レイヤ1の不透明度'},
 
-					{name: 'l2', type: 'str', val: hPrm.l2?.val ?? '(消去)', hint: 'レイヤ2の画像名'},
+					{name: 'l2', type: 'select', val: hPrm.l2?.val ?? '(消去)', hint: 'レイヤ2の画像名', key: '画像ファイル名'},
 					{name: 'f2', type: 'str', val: hPrm.f2?.val ?? '', hint: 'レイヤ2のface属性'},
 					{name: 'pos2', type: 'str', val: hPrm.pos2?.val ?? '', hint: 'レイヤ2のpos位置'},
 					{name: 'o2', type: 'num', val: hPrm.o2?.val ?? '', hint: 'レイヤ2の不透明度'},
@@ -564,40 +626,29 @@ v.type === 'bool' ?'pt-2' :''
 				],
 			};
 		},
-		fg			: hPrm=> {
-			let col = 0;
-			switch (hPrm.layer?.val) {
-				case 'base': col = 1; break;
-				case '0': col = 2; break;
-				case '1': col = 3; break;
-				case '2': col = 4; break;
-				case 'mes': col = 5; break;
-				case 'mes_c2p': col = 6; break;
-			}
-			return {
-				col			: col,
-				btn_style	: `btn-success`,
-				icon		: 'fa-image',
-				btn_face	: hPrm.fn?.val ?? '',
-			};
-		},
+		fg			: hPrm=> {return {
+			col			: hPrm.layer?.val ?? '命令',
+			btn_style	: `btn-success`,
+			icon		: 'fa-image',
+			btn_face	: hPrm.fn?.val ?? '',
+		};},
 
 		button		: hPrm=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-sign-out-alt',
 			btn_face	: (hPrm.text?.val ?`文字ボタン「${hPrm.text?.val}」` :'')
 						+ (hPrm.pic?.val ?`画像ボタン ${hPrm.pic?.val}` :''),
 		}},
 		enable_event: hPrm=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-check-square',
 			btn_face	: `イベント発生 ${(hPrm.enabled?.val === 'true')
 				?'させる' :'させない'}`,
 		}},
 		event		: hPrm=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-bolt',
 			btn_face	: 'イベント予約'
@@ -607,31 +658,31 @@ v.type === 'bool' ?'pt-2' :''
 		}},
 
 		waitclick	: _=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-hourglass-start',
 			btn_face	: 'クリック待ち',
 		}},
 		r			: _=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-align-left',
 			btn_face	: '文字改行',
 		}},
 		l			: _=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-leaf',
 			btn_face	: '改行待ち',
 		}},
 		plc			: hPrm=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-book-open',
 			btn_face	: '改ページ待ち'
 				+ (hPrm.visible?.val === 'false' ?'（マーク非表示）' :''),
 			args		: [
-				{name: 'visible', type: 'bool', val: hPrm.visible?.val ?? 'true', hint: 'マーク表示するか'},
+				{name: 'visible', type: 'bool', val: hPrm.visible?.val ?? 'true', hint: 'マーク表示する'},
 				{name: 'time', type: 'num', val: hPrm.time?.val ?? '500', hint: '再開時の効果音FO時間'},
 				{name: 'buf', type: 'str', val: hPrm.buf?.val ?? '音声', hint: '再開時の再生停止'},
 			],
@@ -644,7 +695,7 @@ v.type === 'bool' ?'pt-2' :''
 		}},
 
 		bgm			: hPrm=> {return {
-			col			: 7,
+			col			: 'BGM',
 			td_style	: 'sn-cmb-'
 				+ (hPrm.fn?.val ?`start" data-fn="${hPrm.fn?.val}` :'end'),
 			btn_style	: 'btn-info',
@@ -653,14 +704,14 @@ v.type === 'bool' ?'pt-2' :''
 			tooltip		: `fn=${hPrm.fn?.val}`,
 		}},
 		stopbgm		: _=> {return {
-			col			: 7,
+			col			: 'BGM',
 			td_style	: 'sn-cmb-end',
 			btn_style	: 'btn-outline-info',
 			icon		: 'fa-stop',
 			btn_face	: '再生停止',
 		}},
 		fadeoutbgm	: hPrm=> {return {
-			col			: 7,
+			col			: 'BGM',
 			td_style	: 'sn-cmb-end',
 			btn_style	: 'btn-outline-info',
 			icon		: 'fa-volume-down',
@@ -668,7 +719,7 @@ v.type === 'bool' ?'pt-2' :''
 		}},
 
 		wait		: hPrm=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-hourglass-start',
 			btn_face	: `${hPrm.time?.val ?? 0}ms`,
@@ -677,7 +728,7 @@ v.type === 'bool' ?'pt-2' :''
 				+ (hPrm.canskip?.val ?` スキップできるか=${hPrm.canskip?.val}` :''),
 		}},
 		s			: _=> {return {
-			col			: 5,
+			col			: 'mes',
 			btn_style	: 'btn-primary',
 			icon		: 'fa-stop',
 			btn_face	: '停止する',
@@ -688,19 +739,19 @@ v.type === 'bool' ?'pt-2' :''
 		}},
 
 		macro	: hPrm=> {
-			CteScore.macro_nm = hPrm.name?.val ?? '';
+			this.macro_nm = hPrm.name?.val ?? '';
 			return {
 				btn_style	: 'btn-outline-secondary text-white',
 				icon		: 'fa-box-open',
-				btn_face	: `マクロ定義の開始 ${CteScore.macro_nm}`,
+				btn_face	: `マクロ定義の開始 ${this.macro_nm}`,
 				args		: [
 					{name: 'name', type: 'str', val: hPrm.name.val ?? '', hint: 'マクロ名'},
 				],
 			}
 		},
 		endmacro		: _=> {
-			const nm = CteScore.macro_nm;
-			CteScore.macro_nm = '';
+			const nm = this.macro_nm;
+			this.macro_nm = '';
 			return {
 				btn_style	: 'btn-outline-secondary text-white',
 				icon		: 'fa-box',
@@ -716,11 +767,24 @@ v.type === 'bool' ?'pt-2' :''
 			],
 		}},
 	};
-	private	static	macro_nm	= '';
+	private	macro_nm	= '';
+	private	MAX_COL		= 7;
+	private	hCn2Col	: {[nm: string]: number}	= {
+		'命令'	: 0,
+		'base'	: 1,
+		'0'		: 2,
+		'1'		: 3,
+		'2'		: 4,
+		'mes'	: 5,
+		'BGM'	: 6,
+		'SE'	: 7,
+		'詳細'	: 8,
+	
+	};
 	private	make_tds(col: number, line: number, btn_style: string, icon: string, btn_face: string, tooltip = '', aft = '', td_style = '', detail = ''): string {
 		return '<td></td>'.repeat(col) +`
 	<td class="p-0 ${td_style}">
-		<button type="button" class="btn btn-block btn-sm ${btn_style}" data-faceicon="${icon}" data-ripple-color="dark" draggable="true"${
+		<button type="button" class="btn btn-block btn-sm ${btn_style}" data-faceicon="${icon}" data-ripple-color="dark" id="${line}btn" draggable="true"${
 			tooltip
 			?` data-mdb-toggle="tooltip" data-placement="right" title="${tooltip}"`
 			:''
@@ -728,11 +792,11 @@ v.type === 'bool' ?'pt-2' :''
 			<i class="fas ${icon}" aria-hidden="true"></i>
 		</button>
 		${aft ?`<div class="dropdown-menu col-xxl-4 col-sm-6 col-12"><form class="p-1 d-flex flex-wrap">${aft}</form></div>` :''}
-	</td>`+
-	'<td></td>'.repeat(8 -col) +`
+	</td>`.replace(' dropdown-toggle', '')+
+	'<td></td>'.repeat(this.MAX_COL -col) +`
 	<td class="p-0">${
 		detail || `
-		<button type="button" class="btn btn-block btn-sm px-2 text-start text-lowercase ${btn_style}" data-faceicon="${icon}" data-ripple-color="dark" id="sn-btn${line}" draggable="true"${
+		<button type="button" class="btn btn-block btn-sm px-2 text-start text-lowercase ${btn_style}" data-faceicon="${icon}" data-ripple-color="dark" id="${line}detail" draggable="true"${
 			tooltip
 			?` data-mdb-toggle="tooltip" data-placement="right" title="${tooltip}"`
 			:''