+++ core/src/ToolBox.ts
@@ -36,7 +36,8 @@ export class ToolBox implements WebviewViewProvider {
 			{icon: 'fa-user', name: '人物登場・切替', style: 'btn-success text-black', scr: '[hu]'},
 			{icon: 'fa-user-slash', name: '人物退場', style: 'btn-outline-success', scr: '[hu_hide]'},
 			{icon: 'fa-running', name: '人物移動', style: 'btn-success text-black', scr: '[hu_mov]'},
-			{icon: 'fa-door-open', name: '場面転換', style: 'btn-success text-black', scr: '[scene_change bg=${bg}]'},
+			{icon: 'fa-door-open', name: '場面転換', style: 'btn-success text-black', scr: '[scene_change bg=${画像レイヤ名:bg} time=1000]'},
+			{icon: 'fa-user-ninja', name: '場面準備', style: 'btn-success btn-rounded text-black', scr: '[scene_reserve layer=0 fn=${画像ファイル名} alpha=1]'},
 		],},
 		{カテゴリ: '音系', 要素: [
 			{icon: 'fa-play', name: 'BGM再生', style: 'btn-info text-black', scr: '[bgm fn=${ＢＧＭファイル名} time=${ms}]'},
@@ -46,7 +47,7 @@ export class ToolBox implements WebviewViewProvider {
 		],},
 		{カテゴリ: '選択肢やジャンプ系', 要素: [
 			{icon: 'fa-project-diagram', name: '選択肢', style: 'btn-primary', scr: ''},
-			{icon: 'fa-tag', name: 'ラベル', style: 'btn-outline-light', scr: '*${ラベル名宣言}'},
+			{icon: 'fa-tag', name: 'ラベル', style: 'btn-outline-light', scr: '*（ラベル名宣言）'},
 			{icon: 'fa-stop', name: '停止', style: 'btn-primary', scr: '[s]'},
 			{icon: 'fa-external-link-alt', name: 'ジャンプ', scr: '[jump fn=${ファイル名} label=${ラベル名}]'},
 			{icon: 'fa-exchange-alt', name: 'コール', scr: '[call fn=${ファイル名} label=${ラベル名}]'},
@@ -54,11 +55,11 @@ export class ToolBox implements WebviewViewProvider {
 		],},
 		{カテゴリ: '演出・その他', 要素: [
 			{icon: '', name: '空行', style: 'btn-outline-primary', scr: '\n\n'},
-			{icon: 'fa-comment-dots', name: 'コメント', style: 'btn-outline-light btn-rounded', scr: ';$$$'},
+			{icon: 'fa-comment-dots', name: 'コメント', style: 'btn-outline-light btn-rounded', scr: ';（コメント）'},
 			{icon: 'fa-hourglass-start', name: '待機', style: 'btn-primary', scr: ''},
 			{icon: 'fa-bolt', name: '振動', scr: '[quake time=${時間ms}]'},
 			{icon: 'fa-th', name: 'アルバム解放', scr: '[アルバム解放 name=$$$]'},
-			{icon: 'fa-calculator', name: '変数操作', style: 'btn-secondary', scr: '[let name=${変数名} text=${文字}]'},
+			{icon: 'fa-calculator', name: '変数操作', style: 'btn-secondary', scr: '[let name=${代入変数名} text=#設定したい値#]'},
 			{icon: 'fa-code', name: 'タグ・マクロ', style: 'btn-light', scr: ''},
 		],},
 	];